+++ app/Http/Controllers/NewsletterController.php
