+++ app/Services/ConvertKitNewsletter.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Services;
+
+class ConvertKitNewsletter implements Newsletter
+{
+    public function subscribe(string $email, string $list = null)
+    {
+        // subscribe the user with ConvertKit-specific
+        // API requests.
+    }
+}
